fix alignment of social Links in desktop
