+++ src/App.jsx
@@ -1,11 +1,12 @@
 import React,{useEffect} from 'react'
-import Hero from './components/Hero/Hero'
-import About from './components/About'
-import Skills from './components/Skills'
-import Works from './components/Works'
-import Contact from './components/Contact'
+import Hero from './Components/Hero/Hero'
+import About from './Components/About'
+import Skills from './Components/Skills'
+import Works from './Components/Portfolio/Works'
+import Contact from './Components/Contact'
 import ReactGA from "react-ga4";
-import Layout from './components/Layout/Layout'
+import Header from './Components/Header/Header'
+import Footer from './Components/Footer/Footer'
 
 const App = () => {
   useEffect(() => {
@@ -13,13 +14,17 @@ const App = () => {
     ReactGA.send({ hitType: "pageview", page: window.location.pathname });
   }, []);
   return (
-    <Layout>
-      <Hero />
-      <About />
-      <Works />
-      <Skills />
-      <Contact />
-    </Layout>
+    <div className="flex flex-col min-h-screen bg-primary-dark-blue">
+      <Header />
+      <main className="w-full overflow-x-hidden flex-grow  pt-20  md:pt-24   md:pl-12 lg:pl-[5%]">
+        <Hero />
+        <About />
+        <Works />
+        <Skills />
+        <Contact />
+      </main>
+      <Footer />
+    </div>
   )
 }
